requirements.txt
streamlit
sounddevice
websockets
load_dotenv